backend/prediction/requirements.txt
tensorflow-cpu==2.20.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
psycopg2-binary
joblib